requirements.txt
click
loguru
prometheus_client
pyyaml
PyInstaller
flask
requests